Title: 자연수 n개의 합, Time: 124ms, Memory: 10MB, Status: Runtime Error - CodeTree

+++ 250204/자연수 n개의 합/sum-of-n-natural-numbers.java
@@ -1,13 +1,13 @@
 import java.util.*;
 public class Main {
-    public static int s;
+    public static long s;
     public static void main(String[] args) {
         // Please write your code here.
         Scanner sc = new Scanner(System.in);
         s = sc.nextInt();
 
         long left = 1;
-        long right = (int)1e9;
+        long right = (int)1e9; // 
         long maxNum = 1;
         while(left <= right){
             long mid = (left + right) / 2;
